+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- CASMCMS-8918: Get SLES packages from `artifactory` instead of `slemaster` to avoid build problems
+
 ## [2.0.0] - 2023-04-05
 ### Changed
  - CASMCMS-8252: Enabled building of unstable artifacts